int16 is an invalid protobuf type xsd:integer and xsd:positiveInteger do not have a maximum values xsd:int is an int32

+++ src/main/java/com/github/tranchis/xsd2thrift/marshal/ProtobufMarshaller.java
@@ -33,9 +33,9 @@ public class ProtobufMarshaller implements IMarshaller
 	public ProtobufMarshaller()
 	{
 		typeMapping = new TreeMap<String,String>();
-		typeMapping.put("positiveInteger", "int16");
+		typeMapping.put("positiveInteger", "int64");
 		typeMapping.put("int", "int32");
-		typeMapping.put("integer", "int32");
+		typeMapping.put("integer", "int64");
 		typeMapping.put("long", "int64");
 		typeMapping.put("decimal", "double");
 		typeMapping.put("ID", "string");
